process location
